refactor: Feature-Sliced Design
